tictactoe: Avoid unnecessary redrawing of all symbols
For every frame generated, the symbols are redrawn for every cell even
if the cell is unchanged. We do not need to need to draw a cell that
has already been changed. Store those cells as a tuple in a list named
"updated_cells" and use that to find what are already drawn and what
more cells need to be drawn later on.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>