tictactoe: Remove an obsolete check_draw() call
When game_over is true, there are two possiblities, one: A player has won,
two: The game is a draw.

If neither player win, that means the game is a draw. Therefore, there's
absolutely no need to check using check_draw() in game_status(). We can
just use a simple "else" and save a function call.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>