tictactoe: Optimize game_initiating_window()
When reset_game() is called in event_handler(), we reinitialize the
dictionary to the default state and call game_initiating_window()
inside reset_game().

And, when the main program is run, the first thing we do is, call
game_initiating_window(). So, why not initialize the dictionary
inside game_initiating_window() to begin with?

This lets us remove reset_game() entirely while simplifying the entire
dictionary reinitialization process.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>