tictactoe: Keep the winning line color dark grey
White for 'O' doesn't suit in my opinion since the 'O' icon is black as
well, like the 'X' icon.

Due to this, remove the use of CIRCLE_COLOR, and rename CROSS_COLOR to
WIN_LINE_COLOR so that we use the same dark grey line for all symbols.

Since there's LINE_COLOR that's used for the color of the grid lines of
the board, it can be misinterpreted. So, rename that to GRID_COLOR.
Rename draw_lines() to draw_grid() as well.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>